# openapi.yaml
openapi: 3.0.3
info:
  title: Pera Agent Wallet API
  version: 0.1.0
  description: "Multi-user REST + MCP API for the Pera dashboard and for AI
    agents. Auth: passkey (WebAuthn) → owner session `ps_…` (everything); scoped
    agent tokens `pat_…` (scopes read | pay | fund | admin) for AI runtimes; the
    static API_BEARER_TOKEN is the admin credential. All amounts are decimal
    USDC strings; every mutation returns tx hashes and emits a per-user event."
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    XdrBody:
      type: object
      properties:
        xdr:
          type: string
          minLength: 20
      required:
        - xdr
      additionalProperties: false
    AuthorizeBuildBody:
      type: object
      properties:
        dailyCapUsdc:
          type: string
          pattern: ^\d+(\.\d{1,7})?$
      additionalProperties: false
...
    CapBody:
      type: object
      properties:
        dailyCapUsdc:
          type: string
          pattern: ^\d+(\.\d{1,7})?$
        weeklyCapUsdc:
          type: string
          pattern: ^\d+(\.\d{1,7})?$
      additionalProperties: false
    RulesBody:
      type: object
      properties:
        weeklyCapUsdc:
          nullable: true
          type: string
          pattern: ^\d+(\.\d{1,7})?$
        maxPerCallUsdc:
          nullable: true
          type: string
          pattern: ^\d+(\.\d{1,7})?$
        allowedNetworks:
          minItems: 1
          type: array
          items:
            type: string
            enum:
              - stellar:testnet
              - eip155:84532
        approveAboveUsdc:
          nullable: true
          type: string
          pattern: ^\d+(\.\d{1,7})?$
      required:
        - weeklyCapUsdc
        - maxPerCallUsdc
        - allowedNetworks
      additionalProperties: false
    RulesApprovalBody:
      type: object
      properties:
        challenge:
          type: string
          minLength: 16
        assertion:
          type: object
          properties:
            id:
              type: string
            rawId:
              type: string
            type:
              type: string
            response:
              type: object
              properties:
                clientDataJSON:
                  type: string
                authenticatorData:
                  type: string
                signature:
                  type: string
                userHandle:
                  nullable: true
                  type: string
              required:
                - clientDataJSON
                - authenticatorData
                - signature
              additionalProperties: false
          required:
            - id
            - response
          additionalProperties: false
      required:
        - challenge
        - assertion
      additionalProperties: false
...
    PayBodyV2:
      type: object
      properties:
        url:
          type: string
          format: uri
        method:
          default: GET
          type: string
          enum:
            - GET
            - POST
            - PUT
            - PATCH
            - DELETE
        headers:
          type: object
          additionalProperties:
            type: string
        body:
          type: string
        prefer:
          default: auto
          type: string
          enum:
            - auto
            - stellar
            - evm
        maxAmountUsdc:
          type: string
          pattern: ^\d+(\.\d{1,7})?$
        approvalId:
          type: string
      required:
        - url
        - method
        - prefer
      additionalProperties: false
    QuoteBody:
      type: object
      properties:
        url:
          type: string
          format: uri
        method:
          default: GET
          type: string
          enum:
            - GET
            - POST
            - PUT
            - PATCH
            - DELETE
        headers:
          type: object
          additionalProperties:
            type: string
        body:
          type: string
        prefer:
          default: auto
          type: string
          enum:
            - auto
            - stellar
            - evm
      required:
        - url
        - method
        - prefer
      additionalProperties: false
...
    EvmTransferBody:
      type: object
      properties:
        to:
          type: string
          pattern: ^0x[0-9a-fA-F]{40}$
        amountUsdc:
          type: string
          pattern: ^\d+(\.\d{1,7})?$
      required:
        - to
        - amountUsdc
      additionalProperties: false
...
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        errorCode:
          type: number
        detail: {}
      required:
        - error
      additionalProperties: false
paths:
...
  /agent/policy:
    get:
      summary: Cap, used, remaining, rule id — read from the spending_limit policy
        contract
      security: *a2
      responses:
        "200":
          description: PolicyView
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    post:
      summary: Submit the passkey-signed cap change (built by /agent/policy/build)
      security: *a2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/XdrBody"
      responses:
        "200":
          description: "{ ruleId, txHash, dailyCapUsdc?, weeklyCapUsdc?, rules? }"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /agent/rules/options:
    post:
      summary: "Propose a ruleset: returns a single-use WebAuthn challenge bound to
        it, to be signed by the passkey that owns the wallet"
      security: *a2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RulesBody"
      responses:
        "200":
          description: "{ challenge, rpId, allowCredentials, rules }"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /agent/rules:
    get:
      summary: Router-enforced agent rules (weekly x402 limit, max price per call,
        allowed chains) + this week's spend
      security: *a2
      responses:
        "200":
          description: "{ weeklyCapUsdc, maxPerCallUsdc, allowedNetworks,
            spentThisWeekUsdc, paymentsThisWeek, enforcedBy }"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    put:
      summary: "Apply the ruleset approved by the owner's passkey: { challenge,
        assertion } from /agent/rules/options. Violations make POST /agent/pay
        answer 409 RULE_VIOLATION before anything is signed"
      security: *a2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RulesApprovalBody"
      responses:
        "200":
          description: "{ weeklyCapUsdc, maxPerCallUsdc, allowedNetworks,
            spentThisWeekUsdc, paymentsThisWeek, enforcedBy }"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400": *a1
  /agent/policy/build:
    post:
      summary: "Build a cap change for passkey signing: daily or weekly window.
        method=set_spending_limit → kit.wallet.fromJSON.execute,
        method=add_policy (weekly window on an older rule) →
        fromJSON.add_policy; then kit.signAdmin"
      security: *a2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CapBody"
      responses:
        "200":
          description: "{ json, xdr, ruleId, window, method, dailyCapUsdc | weeklyCapUsdc }"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /agent/authorize/build:
    post:
      summary: Build add_context_rule (agent signer + spending_limit) for passkey
        signing
      security: *a2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorizeBuildBody"
      responses:
        "200":
          description: "{ json, xdr, agentPublicKey, dailyCapUsdc, ruleName,
            smartAccountId }"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /agent/authorize:
    post:
      summary: Submit the passkey-signed rule transaction sponsored; records the rule id
      security: *a2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/XdrBody"
      responses:
        "200":
          description: "{ ruleId, txHash, explorerUrl }"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /stellar/submit:
    post:
      summary: Sponsored submission of any passkey-signed transaction targeting the
        user's smart account
      security: *a2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/XdrBody"
      responses:
        "200":
          description: "{ hash, explorerUrl, ledger }"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /agent/pay:
    post:
      summary: "Pay a paywalled URL (x402): rules → approval threshold → on-chain cap
        → settlement; returns the paid body, tx hashes and the event timeline.
        Scope: pay"
      security: *a2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayBodyV2"
      responses:
        "200":
          description: PayResult
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "403": *a1
        "409":
          description: RULE_VIOLATION | SPENDING_CAP_EXCEEDED | APPROVAL_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502": *a1
  /agent/pay/over-cap-demo:
    post:
      summary: Fund the smart account then attempt a top-up above the cap; returns the
        on-chain rejection
      security: *a2
      responses:
        "200":
          description: OverCapResult
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "409": *a1
  /evm/transfer:
    post:
      summary: Gasless USDC transfer from the user's EVM wallet (Privy sponsorship or
        EIP-3009 relay)
      security: *a2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvmTransferBody"
      responses:
        "200":
          description: "{ hash, explorerUrl, sponsored }"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /agent/quote:
    post:
      summary: "Probe a paywalled URL without paying: offers (price/network/payee) and
        the rules' verdict (allow | requires_approval | denied). Scope: read"
      security: *a2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteBody"
      responses:
        "200":
          description: Quote + verdict
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "403": *a1
  /agent/services:
    get:
      summary: "Paywalled endpoints the demo resource server advertises. Scope: read"
      security: *a2
      responses:
        "200":
          description: "{ resourceServer, endpoints[] }"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /agent/connect/hermes:
    post:
      summary: Mint a read+pay token and return the Hermes config snippet, env line,
        deep link and Claude Code command (owner session)
      security: *a2
      responses:
        "200":
          description: "{ tokenId, scopes, expiresAt, mcpUrl, snippetYaml, envLine,
            deepLink, claudeCode, stdioShim, skill }"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...